Update to the last version used in NetRush.
Notes: 
 - The interpolation part have to be rewritten. 
 - It might be a good idea to separate the code in several libs...

+++ trunk/GCore/GC_Time.h
@@ -15,7 +15,7 @@ namespace gcore
 	typedef double TimeValue;
 
 	/// Type for time flow factor value.
-	typedef float TimeFlowFactor;
+	typedef double TimeFlowFactor;
 
 }
